cryoem_api/requirements.txt
django
django-cors-headers
djangorestframework
python-dotenv
coverage
coreapi
Pillow